+++ string.c
@@ -45,9 +45,9 @@ shiftstring(int up)
 	struct string *sp;
 	for(sp=string; sp<=string+NSTRING; sp++)
 		if(up)
-			sp->str += (int)strarea;
+			sp->str += (long)strarea;
 		else
-			sp->str -= (int)strarea;
+			sp->str -= (long)strarea;
 }
 
 void
